修复提交订单时出现 Jva NullPointerException 异常

+++ src/main/java/com/xuyang/gdutmallserver/controller/OrderController.java
@@ -1,41 +1,31 @@
 package com.xuyang.gdutmallserver.controller;
 
 import com.xuyang.gdutmallserver.common.InitAction;
-import com.xuyang.gdutmallserver.domain.BaseResp;
-import com.xuyang.gdutmallserver.domain.CancelOrderReq;
-import com.xuyang.gdutmallserver.domain.ConfirmOrderReq;
-import com.xuyang.gdutmallserver.domain.GetOrderByIdReq;
-import com.xuyang.gdutmallserver.domain.GetOrderListReq;
-import com.xuyang.gdutmallserver.domain.PayOrderReq;
-import com.xuyang.gdutmallserver.domain.SubmitOrderReq;
+import com.xuyang.gdutmallserver.domain.*;
 import com.xuyang.gdutmallserver.model.MessageInfo;
 import com.xuyang.gdutmallserver.model.Order;
 import com.xuyang.gdutmallserver.model.OrderGoods;
 import com.xuyang.gdutmallserver.model.OrderInfo;
-import com.xuyang.gdutmallserver.service.CartGoodsService;
-import com.xuyang.gdutmallserver.service.MessageService;
-import com.xuyang.gdutmallserver.service.OrderService;
-import com.xuyang.gdutmallserver.service.ShipAddressService;
-import com.xuyang.gdutmallserver.service.UserService;
+import com.xuyang.gdutmallserver.service.*;
 import com.xuyang.gdutmallserver.utils.DateUtil;
 import com.xuyang.gdutmallserver.utils.YuanFenConverter;
 import com.xuyang.gdutmallserver.utils.push.PushSender;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
 import static com.xuyang.gdutmallserver.controller.BaseController.DEFAULT_JSON_CONTENT_TYPE;
 
 @Controller
 @RequestMapping(produces = {DEFAULT_JSON_CONTENT_TYPE}, value = {"/order"})
 public class OrderController extends BaseController {
+
 
     @Autowired
     private OrderService orderService;
@@ -56,14 +46,14 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp<List<Order>> getOrderList(@RequestBody GetOrderListReq req) {
         BaseResp resp = new BaseResp();
- 
+
         List<OrderInfo> list = this.orderService.getOrderList(Integer.valueOf(this.request.getHeader("token")), req.getOrderStatus());
         if ((list == null) || (list.size() == 0)) {
             resp.setStatus(0);
             resp.setMessage("列表为空");
             return resp;
         }
- 
+
         List orderList = new ArrayList();
         for (OrderInfo info : list) {
             Order order = new Order();
@@ -72,17 +62,17 @@ public class OrderController extends BaseController {
             order.setPayType(info.getPayType());
             order.setTotalPrice(YuanFenConverter.changeY2F(Long.valueOf(info.getTotalPrice())));
             order.setShipAddress(this.shipAddressService.getShipAddressById(info.getShipId()));
- 
+
             List<OrderGoods> goodsList = this.orderService.getOrderGoodsList(info.getId());
             for (OrderGoods orderGoods : goodsList) {
                 orderGoods.setGoodsPrice(YuanFenConverter.changeY2F(orderGoods.getGoodsPrice()));
             }
- 
+
             order.setOrderGoodsList(goodsList);
- 
+
             orderList.add(order);
         }
- 
+
         resp.setStatus(0);
         resp.setMessage("列表获取成功");
         resp.setData(orderList);
@@ -93,24 +83,24 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp<Order> getOrderById(@RequestBody GetOrderByIdReq req) {
         BaseResp resp = new BaseResp();
- 
+
         Order order = new Order();
- 
+
         OrderInfo orderInfo = this.orderService.getOrderById(req.getOrderId());
         order.setId(orderInfo.getId());
         order.setOrderStatus(orderInfo.getOrderStatus());
         order.setPayType(orderInfo.getPayType());
         order.setTotalPrice(YuanFenConverter.changeY2F(Long.valueOf(orderInfo.getTotalPrice())));
- 
+
         order.setShipAddress(this.shipAddressService.getShipAddressById(orderInfo.getShipId()));
- 
+
         List<OrderGoods> orderGoodsList = this.orderService.getOrderGoodsList(orderInfo.getId());
         for (OrderGoods orderGoods : orderGoodsList) {
             orderGoods.setGoodsPrice(YuanFenConverter.changeY2F(orderGoods.getGoodsPrice()));
         }
- 
+
         order.setOrderGoodsList(orderGoodsList);
- 
+
         resp.setStatus(0);
         resp.setMessage("获取订单成功");
         resp.setData(order);
@@ -121,21 +111,24 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp submitOrder(@RequestBody SubmitOrderReq req) {
         BaseResp resp = new BaseResp();
- 
+
         OrderInfo orderInfo = this.orderService.getOrderById(req.getOrder().getId());
         orderInfo.setOrderStatus(Integer.valueOf(1));
         orderInfo.setShipId(req.getOrder().getShipAddress().getId());
- 
+
         this.orderService.modifyOrder(orderInfo);
- 
-        List cartIdList = (List) InitAction.cartIdMap.get(req.getOrder().getId());
-        this.cartGoodsService.deleteCartGoods(cartIdList);
-        InitAction.cartIdMap.remove(req.getOrder().getId());
- 
+
+        if (InitAction.cartIdMap != null) {
+            List cartIdList = InitAction.cartIdMap.get(req.getOrder().getId());
+            this.cartGoodsService.deleteCartGoods(cartIdList);
+            InitAction.cartIdMap.remove(req.getOrder().getId());
+        }
+
+
         int userId = Integer.valueOf(this.request.getHeader("token")).intValue();
         String pushId = this.userService.getUserById(userId).getPushId();
         sendMessage(Integer.valueOf(userId), pushId, req.getOrder().getId() + "");
- 
+
         resp.setStatus(0);
         resp.setMessage("订单提交成功");
         return resp;
@@ -150,7 +143,7 @@ public class OrderController extends BaseController {
         msg.setMsgTime(curTime);
         msg.setUserId(userId);
         this.messageService.addMessage(msg);
- 
+
         PushSender.sendOrderEvent(pushId, orderId);
     }
 
@@ -158,11 +151,11 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp cancelOrder(@RequestBody CancelOrderReq req) {
         BaseResp resp = new BaseResp();
- 
+
         OrderInfo orderInfo = this.orderService.getOrderById(req.getOrderId());
         orderInfo.setOrderStatus(Integer.valueOf(4));
         this.orderService.modifyOrder(orderInfo);
- 
+
         resp.setStatus(0);
         resp.setMessage("订单取消成功");
         return resp;
@@ -172,11 +165,11 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp confirmOrder(@RequestBody ConfirmOrderReq req) {
         BaseResp resp = new BaseResp();
- 
+
         OrderInfo orderInfo = this.orderService.getOrderById(req.getOrderId());
         orderInfo.setOrderStatus(Integer.valueOf(3));
         this.orderService.modifyOrder(orderInfo);
- 
+
         resp.setStatus(0);
         resp.setMessage("订单确认收货成功");
         return resp;
@@ -186,11 +179,11 @@ public class OrderController extends BaseController {
     @ResponseBody
     public BaseResp payOrder(@RequestBody PayOrderReq req) {
         BaseResp resp = new BaseResp();
- 
+
         OrderInfo orderInfo = this.orderService.getOrderById(req.getOrderId());
         orderInfo.setOrderStatus(Integer.valueOf(2));
         this.orderService.modifyOrder(orderInfo);
- 
+
         resp.setStatus(0);
         resp.setMessage("订单支付成功");
         return resp;
